+++ lib/comic_book/cbz.rb
@@ -27,30 +27,14 @@ class ComicBook
       ComicInfo.load xml
     end
 
-    def pages
-      entries = []
+    def entries
+      names = []
 
       Zip::File.open(path) do |zipfile|
-        zipfile.each { |entry| entries << entry.name }
+        zipfile.each { |entry| names << entry.name if entry.file? }
       end
 
-      entries.select { image_file? it }
-             .map    { create_page_from_entry it }
-             .sort_by(&:name)
-    end
-
-    private
-
-    def create_page_from_entry entry
-      basename = File.basename entry
-
-      ComicBook::Page.new entry, basename
-    end
-
-    def image_file? filename
-      extension = File.extname filename.downcase
-
-      ComicBook::IMAGE_EXTENSIONS.include? extension
+      names.map { ComicBook::Entry.new it }
     end
   end
 end